refactor(table): define only the interface for Table not for TableUI
